* commit '3aecf54308d60eea10fe10a5c18e4dff455aef66':
  Use long as file size instead of int while using OPP.

+++ src/com/android/bluetooth/opp/BluetoothOppSendFileInfo.java
@@ -122,7 +122,7 @@ public class BluetoothOppSendFileInfo {
                 try {
                     if (metadataCursor.moveToFirst()) {
                         fileName = metadataCursor.getString(0);
-                        length = metadataCursor.getInt(1);
+                        length = metadataCursor.getLong(1);
                         if (D) Log.d(TAG, "fileName = " + fileName + " length = " + length);
                     }
                 } finally {
