+++ src/com/android/bluetooth/opp/BluetoothOppUtility.java
@@ -75,9 +75,9 @@ public class BluetoothOppUtility {
                 info.mStatus = cursor.getInt(cursor.getColumnIndexOrThrow(BluetoothShare.STATUS));
                 info.mDirection = cursor.getInt(cursor
                         .getColumnIndexOrThrow(BluetoothShare.DIRECTION));
-                info.mTotalBytes = cursor.getInt(cursor
+                info.mTotalBytes = cursor.getLong(cursor
                         .getColumnIndexOrThrow(BluetoothShare.TOTAL_BYTES));
-                info.mCurrentBytes = cursor.getInt(cursor
+                info.mCurrentBytes = cursor.getLong(cursor
                         .getColumnIndexOrThrow(BluetoothShare.CURRENT_BYTES));
                 info.mTimeStamp = cursor.getLong(cursor
                         .getColumnIndexOrThrow(BluetoothShare.TIMESTAMP));
